Move decode function back to main depthai module
